+++ server/component/style/StyleModel.php
@@ -440,13 +440,13 @@ class StyleModel extends BaseModel implements IStyleModel
             $field['content'] = $this->calc_dynamic_values($field, $data_config, $user_name, $user_code);
 
             $default = $field["default_value"] ?? "";
-            if ($field['name'] == "url")
+            if ($field['name'] == "url") {
                 $field['content'] = $this->get_url($field['content']);
-            else if ($field['type'] == "markdown" && (!$this->entry_record || count($this->entry_record) == 0))
-            $field['content'] = $this->parsedown->text($field['content']);
-            else if ($field['type'] == "markdown-inline" && ($this->entry_record && count($this->entry_record) == 0))
-            $field['content'] = $this->parsedown->line($field['content']);
-            else if ($field['type'] == "json") {
+            } else if ($field['type'] == "markdown" && (!$this->entry_record || count($this->entry_record) == 0)) {
+                $field['content'] = $this->parsedown->text($field['content']);
+            } else if ($field['type'] == "markdown-inline" && ($this->entry_record && count($this->entry_record) == 0)) {
+                $field['content'] = $this->parsedown->line($field['content']);
+            } else if ($field['type'] == "json") {
                 $field['content'] = json_decode($field['content'], true);
                 /* $field['content'] = $this->json_style_parse($field['content']); */
             } else if ($field['type'] == "condition") {