+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  ### Bugfix
   - Fix `tabs` loading. Now correctly load the first tab if none is selected. 
   - Fix `simpleMDE` text-area line selection on click.
+  - Fix `help` MDE parse
 
 # v5.0.0